requirements.txt
eventlet==0.23.0
Flask==1.0.2
greenlet==0.4.13
gunicorn==19.8.1
python-socketio==1.9.0
SQLAlchemy==1.2.7
PyYAML==3.13
pyzmail36==1.0.3
